Struct travel::models::repository::todo_repository::MockTodoRepository

pub struct MockTodoRepository {
    TodoRepository_expectations: MockTodoRepository_TodoRepository,
}

Fields
TodoRepository_expectations: MockTodoRepository_TodoRepository
Implementations

impl MockTodoRepository

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockTodoRepository

pub fn expect_find_todo_list_group_by_id(&mut self) -> &mut Expectation

Create an Expectation for mocking the find_todo_list_group_by_id method

pub fn expect_list_todo_list_group(&mut self) -> &mut Expectation

Create an Expectation for mocking the list_todo_list_group method

pub fn expect_save_todo_list_group(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_todo_list_group method

pub fn expect_find_todo_by_id(&mut self) -> &mut Expectation

Create an Expectation for mocking the find_todo_by_id method

pub fn expect_list_todo(&mut self) -> &mut Expectation

Create an Expectation for mocking the list_todo method

pub fn expect_save_todo(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_todo method